requirements.txt
playwright==1.42.0
python-docx==1.1.0
python-dateutil==2.8.2 